requirements.txt
flask
sounddevice
soundfile
openai-whisper
torch
argostranslate
